Make components to depend on observables implicitly
