+++ src/components/__tests__/AeInputAddress.js
@@ -2,7 +2,6 @@ import Vue from 'vue';
 import { mount } from '@vue/test-utils';
 import AeInputAddress from '../AeInputAddress.vue';
 
-jest.mock('../../observables.js');
 Vue.prototype.$globals = {
   IS_MOBILE_DEVICE: true,
 };
