+++ src/components/desktop/Header.vue
@@ -46,7 +46,6 @@ import { mapMutations } from 'vuex';
 import { AeIcon, AeIdenticon } from '@aeternity/aepp-components-3';
 import ButtonPlain from '../ButtonPlain.vue';
 import prefixedAmount from '../../filters/prefixedAmount';
-import { activeAccount } from '../../observables';
 
 export default {
   components: { AeIcon, AeIdenticon, ButtonPlain },
@@ -74,7 +73,9 @@ export default {
       iconName: 'settings',
     }],
   }),
-  subscriptions: () => ({ account: activeAccount }),
+  subscriptions() {
+    return { account: this.$store.state.observables.activeAccount };
+  },
   methods: {
     prefixedAmount,
     ...mapMutations(['toggleSidebar']),
