+++ src/components/mobile/AccountSwitcher.vue
@@ -59,6 +59,7 @@
 
 <script>
 import { mapState, mapMutations } from 'vuex';
+import { pick } from 'lodash-es';
 import { AeIcon } from '@aeternity/aepp-components-3';
 import { directive as clickaway } from 'vue-clickaway';
 import ListItem from '../ListItem.vue';
@@ -66,7 +67,6 @@ import ListItemAccount from '../ListItemAccount.vue';
 import AeCard from '../AeCard.vue';
 import AeRadio from '../AeRadio.vue';
 import Balance from '../Balance.vue';
-import { accounts, totalBalance } from '../../observables';
 
 export default {
   directives: {
@@ -81,7 +81,9 @@ export default {
     Balance,
   },
   computed: mapState(['selectedIdentityIdx']),
-  subscriptions: () => ({ accounts, totalBalance }),
+  subscriptions() {
+    return pick(this.$store.state.observables, ['accounts', 'totalBalance']);
+  },
   methods: mapMutations(['selectIdentity', 'toggleAccountSwitcher']),
 };
 </script>
