+++ src/pages/mobile/Transfer.vue
@@ -115,6 +115,7 @@
 
 <script>
 import BigNumber from 'bignumber.js';
+import { pick } from 'lodash-es';
 import { AeIcon } from '@aeternity/aepp-components-3';
 import moneyWithWingsEmoji from 'emoji-datasource-apple/img/apple/64/1f4b8.png';
 import manTippingHandEmoji from 'emoji-datasource-apple/img/apple/64/1f481-200d-2642-fe0f.png';
@@ -129,7 +130,6 @@ import Menu from '../../components/Menu.vue';
 import MenuItem from '../../components/MenuItem.vue';
 import ListItem from '../../components/ListItem.vue';
 import TransferNotification from '../../components/TransferNotification.vue';
-import { activeAccount } from '../../observables';
 
 export default {
   components: {
@@ -167,7 +167,9 @@ export default {
       showTransferNotification: !!this.transactionHash,
     };
   },
-  subscriptions: () => ({ activeAccount }),
+  subscriptions() {
+    return pick(this.$store.state.observables, ['activeAccount']);
+  },
   mounted() {
     if (this.showTransferNotification) {
       setTimeout(() => { this.showTransferNotification = false; }, 5000);
